refactor(StringConstants) : 문자열 상수 클래스 분리 - FormatString 상수 클래스 추가
